+++ assets/images/Stages/Level-1/Main-1/Main/Spec/Level-1-Main-1-Main-5Spec.c
@@ -21,7 +21,7 @@
 //												DECLARATIONS
 //---------------------------------------------------------------------------------------------------------
 
-extern BYTE Level_1_Main_1_Main_5Map[];
+extern uint16 Level_1_Main_1_Main_5Map[];
 
 extern CharSetROMSpec Level1Main1MainCharset;
