+++ assets/images/Stages/Level-1/Main-1/MainFront/Spec/Level-1-Main-1-MainFront-5Spec.c
@@ -19,7 +19,7 @@
 //												DECLARATIONS
 //---------------------------------------------------------------------------------------------------------
 
-extern BYTE Level_1_Main_1_MainFront_5Map[];
+extern uint16 Level_1_Main_1_MainFront_5Map[];
 extern CharSetROMSpec Level1Main1MainFrontCharset;
