+++ assets/images/Stages/Level-1/Main-1/MainFront/Spec/Level-1-Main-1-MainFront-6Spec.c
@@ -18,7 +18,7 @@
 //												DECLARATIONS
 //---------------------------------------------------------------------------------------------------------
 
-extern BYTE Level_1_Main_1_MainFront_6Map[];
+extern uint16 Level_1_Main_1_MainFront_6Map[];
 extern CharSetROMSpec Level1Main1MainFrontCharset;
